# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "HOTEL",
        "description": "API for hotel management",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/reservations": {
            "get": {
                "operationId": "allReservations",
                "tags": [
                    "Reservations"
                ],
                "description": "Recupere la liste des reservations",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "checkin_date": {
                                        "type": "string"
                                    },
                                    "checkout_date": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "Not found"
                },
                "500": {
                    "description": "Internal server error"
                }
            },
            "post": {
                "operationId": "addReservation",
                "tags": [
                    "Reservations"
                ],
                "description": "Ajoute une reservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_client": {
                                    "type": "integer"
                                },
                                "id_room": {
                                    "type": "integer"
                                },
                                "checkin_date": {
                                    "type": "string"
                                },
                                "checkout_date": {
                                    "type": "string"
                                },
                                "total_price": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "reservation_status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/reservations/{id_reservation}": {
            "get": {
                "operationId": "ReservationsById",
                "tags": [
                    "Reservations"
                ],
                "description": "Recupere la liste des reservations par id de reservation",
                "parameters": [
                    {
                        "name": "id_reservation",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "checkin_date": {
                                        "type": "string"
                                    },
                                    "checkout_date": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "Not found"
                },
                "500": {
                    "description": "Internal server error"
                }
            },
            "patch": {
                "operationId": "updateReservation",
                "tags": [
                    "Reservations"
                ],
                "description": "Met a jour une reservation",
                "parameters": [
                    {
                        "name": "id_reservation",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_client": {
                                    "type": "integer"
                                },
                                "id_room": {
                                    "type": "integer"
                                },
                                "checkin_date": {
                                    "type": "string"
                                },
                                "checkout_date": {
                                    "type": "string"
                                },
                                "total_price": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "reservation_status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "operationId": "deleteReservation",
                "tags": [
                    "Reservations"
                ],
                "description": "Supprime une reservation",
                "parameters": [
                    {
                        "name": "id_reservation",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/reservations/status/{reservation_status}": {
            "get": {
                "operationId": "ReservationsByStatus",
                "tags": [
                    "Reservations"
                ],
                "description": "Recupere la liste des reservations par statut de reservation",
                "parameters": [
                    {
                        "name": "reservation_status",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "checkin_date": {
                                        "type": "string"
                                    },
                                    "checkout_date": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "Not found"
                },
                "500": {
                    "description": "Internal server error"
                }
            }
        },
...
        "/reservations/byRoomType/{roomType}": {
            "get": {
                "operationId": "ReservationsByRoomType",
                "tags": [
                    "Reservations"
                ],
                "description": "Recupere la liste des reservations par type de chambre",
                "parameters": [
                    {
                        "name": "roomType",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "checkin_date": {
                                        "type": "string"
                                    },
                                    "checkout_date": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    },
                                    "room_number": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "Not found"
                },
                "500": {
                    "description": "Internal server error"
                }
            }
        },
...
    }
}